P3 hygiene: clippy-clean, safe Vertex, FFI connectivity exposure, README accuracy
- Remove unnecessary unsafe impl Send/Sync for Vertex (all fields already
  thread-safe; compiler now verifies).
- Clippy: 0 warnings across all targets (was ~58). std::hint::black_box in
  benches, clamp idiom, (0..dim).collect() combination init, public
  BoxedConstraint alias, handle type aliases in ffi, targeted allows in
  examples/tests, doc-indent fix, allow(should_implement_trait) on
  Sobol::next.
- FFI: expose ShgoConnectivityMethod + knn_neighbors through the C API
  (both headers updated; ABI change to ShgoOptions struct); correct the
  f_tol documentation (it gates the f_min criterion, which the C API does
  not expose; default aligned to 1e-4). C/C++ examples rebuilt and run
  against the new headers; committed example binaries removed from git and
  ignored.
- README: connectivity-methods section + option-table rows, softened
  "100% fidelity" claim to reference the documented audit, verification
  section now describes the actual fixture coverage (incl. e2e fixtures).
- Delete stray macOS AppleDouble files.
- Benchmark bitwise identical to the P2 baseline (pure hygiene, no
  behavioral change); 103 unit + 11 cross-validation + 21 doctests green.

+++ examples/difficult_benchmark.rs
@@ -50,7 +50,7 @@ fn noisy_schwefel(x: &[f64]) -> f64 {
     // Simple pseudo-random based on counter and coordinates
     let mut hash = count;
     for &xi in x.iter() {
-        hash ^= (xi.to_bits() as u64).wrapping_mul(0x9e3779b97f4a7c15);
+        hash ^= xi.to_bits().wrapping_mul(0x9e3779b97f4a7c15);
         hash = hash.wrapping_mul(0x517cc1b727220a95);
     }
